+++ packages/material-composer-modules/README.md
@@ -1,0 +1,1 @@
+# @material-composer/modules
