+++ packages/material-composer-modules/src/Billboard.ts
@@ -1,4 +1,4 @@
-import { ModuleFactory } from "."
+import { ModuleFactory } from "material-composer"
 import { Billboard as BillboardUnit } from "shader-composer-toybox"
 
 export const Billboard: ModuleFactory = () => (state) => ({
