+++ packages/material-composer-modules/src/Particles.ts
@@ -1,5 +1,5 @@
 import { $, Input, Vec3 } from "shader-composer"
-import { ModuleFactory } from "./index"
+import { ModuleFactory } from "material-composer"
 
 export const Particles: ModuleFactory<{ Progress: Input<"float"> }> = ({
   Progress
