+++ packages/material-composer-modules/src/Rotate.ts
@@ -1,11 +1,13 @@
+import { ModuleFactory } from "material-composer"
 import { Input, Mul } from "shader-composer"
-import { Module } from "./index"
 
 type RotateProps = {
   rotation: Input<"mat3">
 }
 
-export const Rotate = ({ rotation }: RotateProps): Module => (state) => ({
+export const Rotate: ModuleFactory<RotateProps> = ({ rotation }) => (
+  state
+) => ({
   ...state,
   position: Mul(state.position, rotation)
 })
