+++ packages/material-composer-modules/src/SoftParticles.ts
@@ -1,6 +1,6 @@
+import { ModuleFactory } from "material-composer"
 import { Input, Mul, Unit } from "shader-composer"
 import { Softness } from "shader-composer-toybox"
-import { ModuleFactory } from "./index"
 
 export const SoftParticles: ModuleFactory<{
   softness: Input<"float">
