+++ packages/material-composer-modules/src/Translate.ts
@@ -1,11 +1,13 @@
+import { ModuleFactory } from "material-composer"
 import { Add, Input, InstanceMatrix, mat3, Mul, pipe } from "shader-composer"
-import { Module } from "./index"
 
 type TranslateProps = {
   offset: Input<"vec3">
 }
 
-export const Translate = ({ offset }: TranslateProps): Module => (state) => ({
+export const Translate: ModuleFactory<TranslateProps> = ({ offset }) => (
+  state
+) => ({
   ...state,
   position: pipe(
     offset,
