+++ packages/material-composer-modules/src/index.ts
@@ -1,3 +1,4 @@
+import { ModuleFactory, type Module } from "material-composer"
 import {
   Add,
   Gradient,
@@ -10,33 +11,7 @@ import {
 } from "shader-composer"
 import { PSRDNoise3D } from "shader-composer-toybox"
 import { Color } from "three"
-import { Heat, HeatOptions } from "../units"
-
-export type ModuleState = {
-  position: Input<"vec3">
-  normal: Input<"vec3">
-  color: Input<"vec3">
-  alpha: Input<"float">
-}
-
-/**
- * A Module is a function that accepts a module state as its input and returns a new module state.
- */
-export type Module = (state: ModuleState) => ModuleState
-
-/**
- * A Module Factory is a function that returns a Module.
- */
-export type ModuleFactory<P extends ModuleFactoryProps = {}> = (
-  props: P
-) => Module
-
-export type ModuleFactoryProps = Record<string, any>
-
-/**
- * A Module Pipe is an array of Modules.
- */
-export type ModulePipe = Module[]
+import { Heat, HeatOptions } from "./units"
 
 export * from "./Acceleration"
 export * from "./Billboard"
@@ -62,7 +37,7 @@ export const SetAlpha = ({ alpha }: { alpha: Input<"float"> }): Module => (
   alpha
 })
 
-export const Module = ({ module }: { module: Module }): Module => module
+export const CustomModule = ({ module }: { module: Module }): Module => module
 
 export type LavaProps = HeatOptions & {
   color?: (heat: Input<"float">) => Unit<"vec3">
