+++ packages/material-composer/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./ComposableMaterial"
+export * from "./modules"
