--- packages/material-composer/src/util/makeAttribute.ts
@@ -1,4 +1,0 @@
-import { InstancedBufferAttribute } from "three"
-
-export const makeAttribute = (count: number, itemSize: number) =>
-  new InstancedBufferAttribute(new Float32Array(count * itemSize), itemSize)
