--- packages/material-composer/src/util/pipeModules.ts
@@ -1,5 +1,0 @@
-import { pipe } from "shader-composer"
-import { ModuleState, Module } from "../modules"
-
-export const pipeModules = (initial: ModuleState, ...modules: Module[]) =>
-  pipe(initial, ...(modules as [Module]))
