--- packages/material-composer/units/package.json
@@ -1,4 +1,0 @@
-{
-  "main": "dist/material-composer-units.cjs.js",
-  "module": "dist/material-composer-units.esm.js"
-}
